[mobile]feat: move BookAddFragment to bookAdding Folder
