+++ mobile/app/src/main/java/com/thoughtworks/mobile/ui/fragments/bookAdding/BookAddFragment.kt
@@ -1,4 +1,4 @@
-package com.thoughtworks.mobile.ui.fragments
+package com.thoughtworks.mobile.ui.fragments.bookAdding
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
