+++ mobile/app/src/main/res/navigation/nav_graph_home.xml
@@ -15,6 +15,6 @@
     </fragment>
     <fragment
         android:id="@+id/addingForms"
-        android:name="com.thoughtworks.mobile.ui.fragments.BookAddFragment"
+        android:name="com.thoughtworks.mobile.ui.fragments.bookAdding.BookAddFragment"
         android:label="BookAddFragment" />
 </navigation>
